+++ Sources/KlaviyoCore/APIModels/CreateEventPayload.swift
